jobs/permanent/ppr-discharges-for-mhr/requirements/prod.txt
google-auth
google-cloud-secret-manager
google-cloud-storage
google-crc32c
six
psycopg2-binary
python-dotenv
requests